Move cave stuff to the correct packages
